aletheia-backend/pyproject.toml
[project]
name = "aletheia-backend"
version = "0.1.0"
description = "Misinformation detection API using FastAPI and Mistral AI for text and image analysis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "mistralai>=1.2.5",
    "transformers>=4.46.3",
    "datasets>=3.1.0",
    "pandas>=2.0.3",
    "numpy>=1.24.4",
    "matplotlib>=3.7.5",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "accelerate>=1.0.1",
    "scikit-learn>=1.3.2",
    "google-adk>=1.18.0",
    "litellm>=1.80.5",
    "selenium>=4.36.0",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
]
authors = [
    { name = "Aletheia Team" }
]
keywords = ["misinformation", "detection", "ai", "fastapi", "mistral"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

[project.urls]
Homepage = "https://github.com/yourusername/aletheia-backend"
Documentation = "https://github.com/yourusername/aletheia-backend#readme"
Repository = "https://github.com/yourusername/aletheia-backend"
Issues = "https://github.com/yourusername/aletheia-backend/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
